+++ changelog.md
@@ -3,6 +3,7 @@
 - Minor code refactoring
 - Added the ability to get a map of registered commands
 - Fixed plugin example
+- Fixed copying configs from the plugin
 
 # AvrixLoader v1.0.0